move helper functions into test file
